removed old fragility curve and renamed FragilityCurveRefactored to FragilityCurve
